rename: PasswordEncoder와 SessionConst auth 패키지로 이동 (#8)
