+++ src/main/java/com/example/newsfeed/auth/consts/SessionConst.java
@@ -1,4 +1,4 @@
-package com.example.newsfeed.user.consts;
+package com.example.newsfeed.auth.consts;
 
 public abstract class SessionConst {
     public static final String LOGIN_USER = "LOGIN_USER";
