+++ src/main/java/com/example/newsfeed/user/service/AuthService.java
@@ -2,7 +2,7 @@ package com.example.newsfeed.user.service;
 
 import com.example.newsfeed.exception.CustomException;
 import com.example.newsfeed.exception.ExceptionType;
-import com.example.newsfeed.user.config.PasswordEncoder;
+import com.example.newsfeed.auth.config.PasswordEncoder;
 import com.example.newsfeed.user.domain.User;
 import com.example.newsfeed.user.dto.AuthRequestDto.LoginRequestDto;
 import com.example.newsfeed.user.dto.AuthRequestDto.RegisterRequestDto;
